+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Allow for calling `__broccoliGetInfo__` without argument
 * Add missing `__broccoliFeatures__` object
 
 # 1.0.0